+++ src/moonblade/types/dynamic_value.rs
@@ -36,7 +36,7 @@ use super::DynamicNumber;
 pub enum DynamicValue {
     List(Arc<Vec<DynamicValue>>),
     Map(Arc<HashMap<String, DynamicValue>>),
-    String(Arc<String>),
+    String(Arc<str>),
     Bytes(Arc<BString>),
     Float(f64),
     Integer(i64),
@@ -670,28 +670,28 @@ impl From<&[u8]> for DynamicValue {
 impl From<&str> for DynamicValue {
     #[inline]
     fn from(value: &str) -> Self {
-        DynamicValue::String(Arc::new(value.to_string()))
+        DynamicValue::String(Arc::from(value))
     }
 }
 
 impl From<Cow<'_, str>> for DynamicValue {
     #[inline]
     fn from(value: Cow<str>) -> Self {
-        DynamicValue::String(Arc::new(value.into_owned()))
+        DynamicValue::String(Arc::from(value))
     }
 }
 
 impl From<String> for DynamicValue {
     #[inline]
     fn from(value: String) -> Self {
-        DynamicValue::String(Arc::new(value))
+        DynamicValue::String(Arc::from(value))
     }
 }
 
 impl From<char> for DynamicValue {
     #[inline]
     fn from(value: char) -> Self {
-        DynamicValue::String(Arc::new(value.to_string()))
+        DynamicValue::String(Arc::from(value.to_string()))
     }
 }